Refactor to BlockchainStatusController.ts for consistency
