+++ apps/status-api/src/modules/ControllerModule.ts
@@ -1,6 +1,6 @@
 import { CacheModule, Module } from '@nestjs/common'
 import { ActuatorController } from '@defichain-apps/libs/actuator'
-import { BlockchainController } from '../controllers/BlockchainController'
+import { BlockchainStatusController } from '../controllers/BlockchainStatusController'
 import { WhaleApiClient } from '@defichain/whale-api-client'
 import { ConfigService } from '@nestjs/config'
 
@@ -12,7 +12,7 @@ import { ConfigService } from '@nestjs/config'
     CacheModule.register()
   ],
   controllers: [
-    BlockchainController,
+    BlockchainStatusController,
     ActuatorController
   ],
   providers: [
